+++ get_detailed_players.py
@@ -1,8 +1,8 @@
 from scrapy.crawler import CrawlerProcess
 
-from src.players.detailed_players_spider import DetailedPlayersSpider, create_db
+from src.players.detailed_players_spider import DetailedPlayersPlayerBaseSpider, create_db
 
 create_db()
 process = CrawlerProcess()
-process.crawl(DetailedPlayersSpider)
+process.crawl(DetailedPlayersPlayerBaseSpider)
 process.start()
