+++ src/items/debut_player_item.py
@@ -1,9 +1,11 @@
 from scrapy.item import Item, Field
+from dataclasses import dataclass
 
-class DebutPlayerItem(Item):
-    name = Field()
-    link = Field()
-    age_at_debut = Field()
-    debut_score = Field()
-    debut_game_outcome = Field()
-    debut_date = Field()
+@dataclass
+class DebutPlayerItem:
+    name: str = None
+    link: str = None
+    age_at_debut: str = None
+    debut_score: str = None
+    debut_game_outcome: str = None
+    debut_date: str = None
