+++ src/players/players_spider.py
@@ -1,6 +1,6 @@
 import json
 from scrapy.crawler import CrawlerProcess
-from src.base_spider import BaseTransfermarktSpider
+from src.base.base_spider import BaseTransfermarktSpider
 import random
 
 user_agent = ''.join((random.choice('abcdefghijklmnopqrstuvwxyz1234567890@') for i in range(10)))
